projeto tudo funcionando

+++ src/Services/schedules/deleteSchedule.service.ts
@@ -1,11 +1,16 @@
 import { AppDataSource } from "../../data-source"
 import { Schedule } from "../../Entities/schedules.entity"
+import { AppError } from "../../Error/appError"
 
 const deleteScheduleService = async (id: string): Promise<boolean> => {
     const scheduleRepository = AppDataSource.getRepository(Schedule)
     const schedule = await scheduleRepository.findOneBy({
         id
     })
+
+    if (!schedule){
+        throw new AppError(404, "Agendamento não encontrado ou desmarcado")
+    }
 
     await scheduleRepository.delete({id})
